Update 6
No changes made to the web application.

The desktop application had the following changes:

- Changed the theme to be a darker one, this will help they eyes of those officers who are working
on night shift adjust better to the dark.

- Added a MenuStrip for each form except the login form.

- The 'exit' on the menu strip now exits the application

- Edited to GetInformation method in the Methods class to support null values instead of giving out an error.

- Added IsDriver method to the Methods class. This will check whether a person holds a Driver's Licence using his ID.

- Removed unnecessary methods.

- The civilian form now is now fully implemented with the search. It also displays a picture of the civilian which is
located in the resources of the project.

- Implemented the vehicle form which is very similar to the civilian form but instead searches and gives out information according
to a licence plate.

- Added licence form which is currently still empty.

- Everything in the Civilian and Vehicle forms but the search are in a panel. Panel is only displayed when a result is found
successfully, otherwise it is hidden and gives out a message stating that no results were found.

- Added a button in both Civilian and Vehicle forms which its function is yet to be implemented. For now, the button shows only
if the user is a driver. In the vehicle form, this button always shows since in order to have a registered vehicle, one must own
a driver's licence.